\name{ChenEtAlList}
\alias{ChenEtAlList}
\docType{data}
\title{
  ChenEtAlList
}
\description{
This is the file containing crossreactive probes.
}
\usage{data("ChenEtAlList")}
\format{
  A data frame with 29233 observations on the following variable.
}
\details{
asdf
}
\source{
asdf
}
\references{
asdf
}
\examples{
data(ChenEtAlList)
## maybe str(ChenEtAlList) ; plot(ChenEtAlList) ...
}
\keyword{datasets}
