\name{PidsleyCrossReactiveProbesEPIC}
\alias{PidsleyCrossReactiveProbesEPIC}
\docType{data}
\title{
  PidsleyCrossReactiveProbesEPIC
}
\description{
This is the file containing crossreactive probes.
}
\usage{data("PidsleyCrossReactiveProbesEPIC")}
\format{
  A data frame with 43254 observations on the following variable.
}
\details{
asdf
}
\source{
asdf
}
\references{
asdf
}
\examples{
data(PidsleyCrossReactiveProbesEPIC)
## maybe str(PidsleyCrossReactiveProbesEPIC) ; plot(PidsleyCrossReactiveProbesEPIC) ...
}
\keyword{datasets}
