\name{non-specific-probes-Illumina27k}
\alias{non-specific-probes-Illumina27k}
\docType{data}
\title{
  non-specific-probes-Illumina27k
}
\description{
This is the file containing crossreactive probes.
}
\usage{data("non-specific-probes-Illumina27k")}
\format{
  A data frame with 0 observations on the following 2 variables.
}
\details{
asdf
}
\source{
asdf
}
\references{
asdf
}
\examples{
data(non-specific-probes-Illumina27k)
## maybe str(non-specific-probes-Illumina27k) ; plot(non-specific-probes-Illumina27k) ...
}
\keyword{datasets}
